+++ app/Http/Controllers/CampaignsBeneficiariesController.php
@@ -16,7 +16,7 @@ class CampaignsBeneficiariesController extends Controller
     public function index()
     {
         // Eloquent
-        // $this->authorize('viewAny' , CampaignsBeneficiaries::class);
+        $this->authorize('viewAny' , CampaignsBeneficiaries::class);
         $campaignsBeneficiaries = CampaignsBeneficiaries::all();
         return new CampaignsBeneficiariesResourceCollection($campaignsBeneficiaries);
     }
@@ -26,7 +26,7 @@ class CampaignsBeneficiariesController extends Controller
      */
     public function store(Request $request)
     {
-        // $this->authorize('create' , CampaignsBeneficiaries::class);
+        $this->authorize('create' , CampaignsBeneficiaries::class);
         // validate
         $valid = Validator($request->all(), [
             'amount' => 'required|numeric|min:0',
@@ -61,7 +61,7 @@ class CampaignsBeneficiariesController extends Controller
      */
     public function show(CampaignsBeneficiaries $CampaignsBeneficiary)
     {
-        // $this->authorize('view' , $CampaignsBeneficiary);
+        $this->authorize('view' , $CampaignsBeneficiary);
         return new CampaignsBeneficiariesResource($CampaignsBeneficiary);
     }
 
@@ -83,7 +83,7 @@ class CampaignsBeneficiariesController extends Controller
         if (!$valid->fails()) {
             // Eloquent
             $campaignsBeneficiaries = CampaignsBeneficiaries::findOrFail($id);
-            // $this->authorize('update' , $campaignsBeneficiaries);
+            $this->authorize('update' , $campaignsBeneficiaries);
             $campaignsBeneficiaries->amount = $request->input('amount');
             $campaignsBeneficiaries->description = $request->input('description');
             $campaignsBeneficiaries->status = $request->input('status');
@@ -106,7 +106,7 @@ class CampaignsBeneficiariesController extends Controller
     {
         // Eloquent
         $campaignsBeneficiaries = CampaignsBeneficiaries::findOrFail($id);
-        // $this->authorize('delete' , $campaignsBeneficiaries);
+        $this->authorize('delete' , $campaignsBeneficiaries);
         $deleted = $campaignsBeneficiaries->delete();
         return new Response(
             [
@@ -119,7 +119,7 @@ class CampaignsBeneficiariesController extends Controller
     {
         //
         $object = CampaignsBeneficiaries::onlyTrashed()->findOrFail($id);
-        // $this->authorize('restore' , $object);
+        $this->authorize('restore' , $object);
         $restored = $object->restore();
         return response()->json(['status' => $restored]);
     }
@@ -128,7 +128,7 @@ class CampaignsBeneficiariesController extends Controller
     {
         //
         $object = CampaignsBeneficiaries::withTrashed()->findOrFail($id);
-        // $this->authorize('forceDelete' , $object);
+        $this->authorize('forceDelete' , $object);
         $deleted = $object->forceDelete();
         // if ($deleted && $object->image) {
         //     // Storage::delete($object->image);
@@ -140,7 +140,7 @@ class CampaignsBeneficiariesController extends Controller
 
     public function Archives()
     {
-        // $this->authorize('restore', $supplier);
+        $this->authorize('restore', $supplier);
         $campaignBeneficiaries = CampaignsBeneficiaries::onlyTrashed()->get();
         $data = CampaignsBeneficiariesResource::collection($campaignBeneficiaries);
         return response()->json(['status' => true, 'message' => 'success', 'data' => $data], 200);
