+++ app/Http/Controllers/CategoryController.php
@@ -17,7 +17,7 @@ class CategoryController extends Controller
     public function index()
     {
         // Eloquent
-        // $this->authorize('viewAny' , Category::class);
+        $this->authorize('viewAny' , Category::class);
         $category = Category::paginate(10); //Pagination
         return new CategoryResourceCollection($category);
     }
@@ -27,7 +27,7 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        // $this->authorize('create' , Category::class);
+        $this->authorize('create' , Category::class);
         // validate
         $valid = validator($request->all(), [
             'name' => 'required|string|min:2|max:45|unique:categories,id',
@@ -55,7 +55,7 @@ class CategoryController extends Controller
      */
     public function show(Category $category)
     {
-        // $this->authorize('view' , $category);
+        $this->authorize('view' , $category);
         return new CategoryResource($category);
     }
 
@@ -75,7 +75,7 @@ class CategoryController extends Controller
         if (!$valid->fails()) {
             // Eloquent
             $category = Category::findOrFail($id);
-            // $this->authorize('update' , $category);
+            $this->authorize('update' , $category);
             $category->name = $request->input('name');
             $category->description = $request->input('description');
             $update = $category->save();
@@ -96,7 +96,7 @@ class CategoryController extends Controller
     {
         // Eloquent
         $category = Category::findOrFail($id);
-        // $this->authorize('delete' , $category);
+        $this->authorize('delete' , $category);
         $deleted = $category->delete();
         return new Response(
             [
@@ -108,7 +108,7 @@ class CategoryController extends Controller
 
     public function Archives()
     {
-        // $this->authorize('restore', $supplier);
+        $this->authorize('restore', $supplier);
         $category = Category::onlyTrashed()->get();
         $data = CategoryResource::collection($category);
         return response()->json(['status' => true, 'message' => 'success', 'data' => $data], 200);
@@ -121,7 +121,7 @@ class CategoryController extends Controller
     {
         //
         $category = Category::onlyTrashed()->findOrFail($id);
-        // $this->authorize('restore' , $category);
+        $this->authorize('restore' , $category);
         $restored = $category->restore();
         return new Response(['status' => $restored]);
     }
@@ -133,7 +133,7 @@ class CategoryController extends Controller
     {
         //
         $category = Category::withTrashed()->findOrFail($id);
-        // $this->authorize('forceDelete' , $category);
+        $this->authorize('forceDelete' , $category);
         $deleted = $category->forceDelete();
         return new Response(['status' => $deleted]);
     }
