+++ app/Http/Controllers/FaqController.php
@@ -22,7 +22,7 @@ class FaqController extends Controller
     public function index()
     {
 
-        // $this->authorize('viewAny', Faq::class);
+        $this->authorize('viewAny', Faq::class);
 
         $faqs = Faq::simplePaginate(10);
         $data = FaqResource::collection($faqs);
@@ -36,7 +36,7 @@ class FaqController extends Controller
     public function store(Request $request)
     {
 
-      //  $this->authorize('create', Faq::class);
+       $this->authorize('create', Faq::class);
         $validator = validator($request->all(), [
             "question" => "string|required|min:5|max:100",
             "answer" => "string|required|min:5|max:250",
@@ -59,7 +59,7 @@ class FaqController extends Controller
      */
     public function show(Faq $faq)
     {
-        // $this->authorize('view', $faq);
+        $this->authorize('view', $faq);
         // return new Response(["status" => true, "data" => $FAQ], Response::HTTP_OK);
         return new FAQResource($faq);
     }
@@ -70,7 +70,7 @@ class FaqController extends Controller
      */
     public function update(Request $request, Faq $faq)
     {
-        // $this->authorize('update', $faq);
+        $this->authorize('update', $faq);
 
         $validator = validator($request->all(), [
             "question" => "string|required|min:5|max:100",
@@ -100,7 +100,7 @@ class FaqController extends Controller
      */
     public function destroy(Faq $faq)
     {
-        //$this->authorize('delete', $faq);
+        $this->authorize('delete', $faq);
 
         $deleted = $faq->delete();
         return new Response(["status" => $deleted, "message" => "successfully deleted"], Response::HTTP_OK);
@@ -108,7 +108,7 @@ class FaqController extends Controller
 
     public function Archives()
     {
-        // $this->authorize('restore', $supplier);
+        $this->authorize('restore', $supplier);
         $Faq = Faq::onlyTrashed()->get();
         $data = FaqResource::collection($Faq);
         return response()->json(['status' => true, 'message' => 'success', 'data' => $data] , 200);
@@ -133,7 +133,7 @@ class FaqController extends Controller
     public function forceDelete(Request $request, $id): Response
     {
         $faq = Faq::withTrashed()->findOrFail($id);
-        // $this->authorize('forceDelete', $faq);
+        $this->authorize('forceDelete', $faq);
 
         $deleted = $faq->forceDelete();
         return new Response(['status' => $deleted]);
